Fixes for arrays of features for users and items

+++ src/main/java/prediction/response/time/message/MahoutCollaborativeFilteringResponseTimePredictor.java
@@ -128,7 +128,7 @@ public abstract class MahoutCollaborativeFilteringResponseTimePredictor<Collabor
 		if (result instanceof String) {
 			prediction = Double.parseDouble((String) result);
 		} else if (result != null) {
-			prediction = (double) result;
+			prediction = (double) (float) result;
 		}
 		
 		return new ResponseTimeRange(prediction, null);
